requirements.txt
Flask
Flask-Cors
Flask-MySQLdb
python-dotenv
google-cloud-storage
pandas
